test: inline fake_mount.go fixture into fake_mount_test.go
fake_mount.go was only used by tests but lived in the production
package, adding to the compiled binary and muddying the API surface.
Move it into the _test.go file so it is only compiled during testing.
