--- pkg/azurelustre/fake_mount.go
@@ -1,76 +1,0 @@
-/*
-Copyright 2020 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package azurelustre
-
-import (
-	"fmt"
-	"strings"
-	"time"
-
-	mount "k8s.io/mount-utils"
-)
-
-type fakeMounter struct {
-	mount.FakeMounter
-}
-
-// Mount overrides mount.FakeMounter.Mount.
-func (f *fakeMounter) Mount(source, target, fstype string, options []string) error {
-	if strings.Contains(source, "ut-container") {
-		return fmt.Errorf("fake Mount: source error")
-	} else if strings.Contains(target, "error_mount") {
-		return fmt.Errorf("fake Mount: target error")
-	}
-
-	return f.FakeMounter.Mount(source, target, fstype, options)
-}
-
-// MountSensitive overrides mount.FakeMounter.MountSensitive.
-func (f *fakeMounter) MountSensitive(source, target, fstype string, options, sensitiveOptions []string) error {
-	if strings.Contains(source, "ut-container-sens") {
-		return fmt.Errorf("fake MountSensitive: source error")
-	} else if strings.Contains(target, "error_mount_sens") {
-		return fmt.Errorf("fake MountSensitive: target error")
-	}
-
-	return f.FakeMounter.MountSensitive(source, target, fstype, options, sensitiveOptions)
-}
-
-// MountSensitiveWithoutSystemdWithMountFlags overrides mount.FakeMounter.MountSensitiveWithoutSystemdWithMountFlags.
-func (f *fakeMounter) MountSensitiveWithoutSystemdWithMountFlags(source, target, fstype string, options, sensitiveOptions, mountFlags []string) error {
-	if strings.Contains(source, "ut-container-sens-mountflags") {
-		return fmt.Errorf("fake MountSensitiveWithoutSystemdWithMountFlags: source error")
-	} else if strings.Contains(target, "error_mount_sens_mountflags") {
-		return fmt.Errorf("fake MountSensitiveWithoutSystemdWithMountFlags: target error")
-	}
-
-	return f.FakeMounter.MountSensitiveWithoutSystemdWithMountFlags(source, target, fstype, options, sensitiveOptions, mountFlags)
-}
-
-func (f *fakeMounter) IsLikelyNotMountPoint(file string) (bool, error) {
-	if strings.Contains(file, "error_is_likely") {
-		return false, fmt.Errorf("fake IsLikelyNotMountPoint: fake error")
-	}
-	if strings.Contains(file, "false_is_likely") {
-		return false, nil
-	}
-	return true, nil
-}
-
-func (f *fakeMounter) UnmountWithForce(target string, _ time.Duration) error {
-	return f.Unmount(target)
-}
